refactored getting tooltips into ClassCluster
