+++ client/src/util/constants.js
@@ -1,2 +1,2 @@
-export const iconWidth = 30;
+export const iconWidth = 25;
 export const marginSize = 7;
